a11y(combobox): add support for describedby and labelledby and improuve external fields handling
