+++ app/javascript/components/ComboAnnuaireEducationSearch.jsx
@@ -4,12 +4,10 @@ import { QueryClientProvider } from 'react-query';
 import ComboSearch from './ComboSearch';
 import { queryClient } from './shared/queryClient';
 
-function ComboAnnuaireEducationSearch(params) {
+function ComboAnnuaireEducationSearch(props) {
   return (
     <QueryClientProvider client={queryClient}>
       <ComboSearch
-        required={params.mandatory}
-        hiddenFieldId={params.hiddenFieldId}
         scope="annuaire-education"
         minimumInputLength={3}
         transformResults={(_, { records }) => records}
@@ -20,6 +18,7 @@ function ComboAnnuaireEducationSearch(params) {
             nom_commune
           }
         }) => [id, `${nom_etablissement}, ${nom_commune} (${id})`]}
+        {...props}
       />
     </QueryClientProvider>
   );
