+++ app/javascript/components/ComboRegionsSearch.jsx
@@ -4,15 +4,14 @@ import { QueryClientProvider } from 'react-query';
 import ComboSearch from './ComboSearch';
 import { queryClient } from './shared/queryClient';
 
-function ComboRegionsSearch(params) {
+function ComboRegionsSearch(props) {
   return (
     <QueryClientProvider client={queryClient}>
       <ComboSearch
-        required={params.mandatory}
-        hiddenFieldId={params.hiddenFieldId}
         scope="regions"
         minimumInputLength={0}
         transformResult={({ code, nom }) => [code, nom]}
+        {...props}
       />
     </QueryClientProvider>
   );
